Empty/requirements.txt
PySide6
opencv-python
pyaudio
google-genai
python-dotenv
pillow
numpy
websockets
argparse
